+++ app/models/kejadian_keamanan.go
@@ -6,7 +6,7 @@ import (
 )
 
 type KejadianKeamanan struct {
-	IdKejadianKeamanan uint8         `json:"id_kejadian_keamanan" gorm:"primary_key;column:id_kejadian_keamanan"`
+	IdKejadianKeamanan int64         `json:"id_kejadian_keamanan" gorm:"primary_key;column:id_kejadian_keamanan"`
 	Tanggal            carbon.Date   `json:"tanggal" gorm:"default:not null"`
 	JenisKejadianId    string        `json:"jenis_kejadian_id" gorm:"default:not null;column:jenis_kejadian_id"`
 	JenisKejadian      JenisKejadian `json:"jenis_kejadian" gorm:"foreign_key:JenisKejadianId;references:IDJenisKejadian"`
