+++ app/models/laporan_keamanan.go
@@ -5,7 +5,7 @@ import (
 )
 
 type LaporanKeamanan struct {
-	IDLaporanKeamanan  uint8            `json:"id_laporan_keamanan" gorm:"primary_key" column:"id_laporan_keamanan"`
+	IDLaporanKeamanan  int64            `json:"id_laporan_keamanan" gorm:"primary_key" column:"id_laporan_keamanan"`
 	LaporanID          int              `json:"laporan_id" gorm:"default: not null" column:"laporan_id"`
 	Laporan            Laporan          `gorm:"foreign_key:LaporanID"`
 	KejadianKeamananID int              `json:"kejadian_keamanan_id" gorm:"default: not null" column:"kejadian_keamanan_id"`
